Revert "Refactor invitations_spec"
This reverts commit 2a236c178803d37f1f71989363773d0e15ba6f6f.
